move path finding out from commands dir
