+++ jetbase/commands/upgrade.py
@@ -1,6 +1,5 @@
 import os
 
-from jetbase.commands.directories import get_migrations_directory
 from jetbase.constants import MIGRATIONS_DIR
 from jetbase.engine.dry_run import process_dry_run
 from jetbase.engine.file_parser import parse_upgrade_statements
@@ -16,6 +15,7 @@ from jetbase.engine.version import (
 from jetbase.enums import MigrationDirectionType, MigrationType
 from jetbase.logging import logger
 from jetbase.models import MigrationRecord
+from jetbase.paths import get_migrations_directory
 from jetbase.repositories.lock_repo import create_lock_table_if_not_exists
 from jetbase.repositories.migrations_repo import (
     create_migrations_table_if_not_exists,
